Oops, fix a merge error

+++ vtb.h
@@ -217,12 +217,12 @@ VPRAGMA_WARNING_POP \
 
 
 
-inline long vmin(long a, long b)
+inline int vmin(int a, int b)
 {
 	return (a < b) ? a : b;
 }
 
-inline long vmax(long a, long b)
+inline int vmax(int a, int b)
 {
 	return (a > b) ? a : b;
 }
